+++ src/routes/accounts/index.ts
@@ -1,0 +1,4 @@
+export * from './signup';
+export * from './users';
+export * from './login';
+export * from './authenticated';
