+++ src/routes/accounts/login.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from 'express';
 import { body } from 'express-validator';
 import jwt from 'jsonwebtoken';
-import { User } from '../models';
-import { PasswordManager } from '../util/hash/index';
+import { User } from '../../models';
+import { PasswordManager } from '../../util/hash/index';
 
 const loginRouter = require('express').Router();
