+++ src/routes/accounts/users.ts
@@ -1,5 +1,5 @@
 import express, { Response } from 'express';
-import { User } from '../models/user';
+import { User } from '../../models/user';
 
 const router = express.Router();
