+++ src/routes/index.ts
@@ -1,4 +1,1 @@
-export * from './signup';
-export * from './users';
-export * from './login';
-export * from './authenticated';
+export * from './accounts';
